+++ packages/@raid-ui/core/theme/src/borders.js
@@ -1,21 +1,19 @@
 
-import { darkAlpha, gray } from './colours'
+import { mapObject } from './utils'
+import { lightAlpha, darkAlpha, gray } from './colours'
 
 const scale = [
   0, 1, 2, 4, 6, 8, 16, 24, 32
 ]
 
+const baseBorderSize = '2px'
+
+const makeBorder = colour => `${baseBorderSize} solid ${colour}`
+
 export const borders = {
-  control: `2px solid ${gray[500]}`,
-  lightAlpha: `1px solid ${darkAlpha[300]}`,
-  light: {
-    100: `1px solid ${gray[200]}`,
-    200: `1px solid ${gray[300]}`
-  },
-  dark: {
-    100: `1px solid ${gray[800]}`,
-    200: `1px solid ${gray[700]}`
-  }
+  lighten: mapObject(lightAlpha, makeBorder),
+  darken: mapObject(darkAlpha, makeBorder),
+  base: mapObject(gray, makeBorder)
 }
 
 export const radii = {